Make instantiation of AndroidFiles overridable
