+++ backends/gdx-backend-android/src/com/badlogic/gdx/backends/android/AndroidApplication.java
@@ -120,8 +120,7 @@ public class AndroidApplication extends Activity implements AndroidApplicationBa
 			: config.resolutionStrategy);
 		input = createInput(this, this, graphics.view, config);
 		audio = createAudio(this, config);
-		this.getFilesDir(); // workaround for Android bug #10515463
-		files = new AndroidFiles(this.getAssets(), this);
+		files = createFiles();
 		net = new AndroidNet(this, config);
 		this.listener = listener;
 		this.handler = new Handler();
@@ -502,5 +501,10 @@ public class AndroidApplication extends Activity implements AndroidApplicationBa
 	@Override
 	public AndroidInput createInput (Application activity, Context context, Object view, AndroidApplicationConfiguration config) {
 		return new DefaultAndroidInput(this, this, graphics.view, config);
+	}
+
+	protected AndroidFiles createFiles() {
+		this.getFilesDir(); // workaround for Android bug #10515463
+		return new DefaultAndroidFiles(this.getAssets(), this);
 	}
 }
