+++ backends/gdx-backend-android/src/com/badlogic/gdx/backends/android/AndroidLiveWallpaper.java
@@ -74,10 +74,7 @@ public class AndroidLiveWallpaper implements AndroidApplicationBase {
 		// input = new AndroidInput(this, this.getService(), null, config);
 
 		audio = createAudio(this.getService(), config);
-
-		// added initialization of android local storage: /data/data/<app package>/files/
-		this.getService().getFilesDir(); // workaround for Android bug #10515463
-		files = new AndroidFiles(this.getService().getAssets(), this.getService());
+		files = createFiles();
 		net = new AndroidNet(this, config);
 		this.listener = listener;
 		clipboard = new AndroidClipboard(this.getService());
@@ -357,6 +354,12 @@ public class AndroidLiveWallpaper implements AndroidApplicationBase {
 	@Override
 	public AndroidInput createInput (Application activity, Context context, Object view, AndroidApplicationConfiguration config) {
 		return new DefaultAndroidInput(this, this.getService(), graphics.view, config);
+	}
+
+	protected AndroidFiles createFiles() {
+		// added initialization of android local storage: /data/data/<app package>/files/
+		this.getService().getFilesDir(); // workaround for Android bug #10515463
+		return new DefaultAndroidFiles(this.getService().getAssets(), this.getService());
 	}
 
 	@Override
